+++ Core/Abstractions/FileDescriptor.cs
@@ -15,6 +15,9 @@ namespace Unity.Cloud.Common
 
         /// <inheritdoc cref="ProjectDescriptor.ProjectId"/>
         public ProjectId ProjectId => DatasetDescriptor.ProjectId;
+        
+        /// <inheritdoc cref="AssetDescriptor.AssetLibraryId"/>
+        public AssetLibraryId AssetLibraryId => DatasetDescriptor.AssetLibraryId;
 
         /// <inheritdoc cref="AssetDescriptor.AssetId"/>
         public AssetId AssetId => DatasetDescriptor.AssetId;